+++ core-client/src/main/scala/de/welt/contentapi/core/client/services/contentapi/ContentSearchService.scala
@@ -1,18 +1,18 @@
 package de.welt.contentapi.core.client.services.contentapi
 
 import javax.inject.{Inject, Singleton}
-
 import com.google.inject.ImplementedBy
 import com.kenshoo.play.metrics.Metrics
 import de.welt.contentapi.core.client.models.ApiContentSearch
 import de.welt.contentapi.core.client.services.CapiExecutionContext
 import de.welt.contentapi.core.client.services.http.RequestHeaders
-import de.welt.contentapi.core.models.{ApiContent, ApiSearchResponse}
+import de.welt.contentapi.core.models.{ApiBatchResult, ApiContent, ApiSearchResponse}
 import play.api.Configuration
 import play.api.libs.json.{JsLookupResult, JsResult}
-import play.api.libs.ws.WSClient
+import play.api.libs.ws.{WSClient, WSResponse}
 
 import scala.concurrent.Future
+import scala.util.Try
 
 /**
   * Reusable (Playframework) Service for Content Search against our search API.
@@ -59,16 +59,17 @@ class ContentSearchServiceImpl @Inject()(ws: WSClient, metrics: Metrics, configu
   extends AbstractService[ApiSearchResponse](ws, metrics, configuration,"search", capi) with ContentSearchService {
 
   import de.welt.contentapi.core.models.ApiReads.apiSearchResponseReads
+  import AbstractService.implicitConversions._
 
-  override val jsonValidate: (JsLookupResult) ⇒ JsResult[ApiSearchResponse] = json ⇒ (json \ "response").validate[ApiSearchResponse]
+  override val validate: WSResponse ⇒ Try[ApiSearchResponse] = response ⇒ (response.json.result \ "response").validate[ApiSearchResponse]
 
   override def search(apiContentSearch: ApiContentSearch)
                      (implicit requestHeaders: RequestHeaders = Seq.empty): Future[ApiSearchResponse] = {
-    super.get(parameters = apiContentSearch.getAllParamsUnwrapped)
+    super.execute(parameters = apiContentSearch.getAllParamsUnwrapped)
   }
 
   override def batchGetForId(ids: Seq[String])(implicit requestHeaders: Option[RequestHeaders]): Future[Seq[ApiContent]] =
     super
-      .get(urlArguments = Nil, parameters = Seq("id" → ids.mkString("|")))
+      .execute(urlArguments = Nil, parameters = Seq("id" → ids.mkString("|")))
       .map(_.results)
 }