+++ core-client/src/test/scala/de/welt/contentapi/core/client/services/contentapi/CircuitBreakerSpec.scala
@@ -4,17 +4,21 @@ import akka.pattern.CircuitBreakerOpenException
 import com.codahale.metrics.Timer.Context
 import de.welt.contentapi.core.client.TestExecutionContext
 import de.welt.contentapi.core.client.services.exceptions.{HttpClientErrorException, HttpRedirectException, HttpServerErrorException}
+import org.mockito.Matchers
+import org.mockito.Matchers.anyString
 import org.mockito.Mockito.{never, times, verify, when}
 import org.scalatest.concurrent.Eventually
 import org.scalatest.mockito.MockitoSugar
 import org.scalatest.time.{Millis, Seconds, Span}
 import org.scalatestplus.play.PlaySpec
 import play.api.Configuration
-import play.api.libs.json.{JsLookupResult, JsResult, JsString}
+import play.api.libs.json.JsString
+import play.api.libs.ws.WSResponse
 import play.api.test.Helpers
 
 import scala.concurrent.Await
 import scala.concurrent.duration._
+import scala.util.Try
 
 //noinspection ScalaStyle
 class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionContext {
@@ -26,7 +30,10 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         class TestService extends AbstractService[String](mockWsClient, metricsMock,
           CircuitBreakerSpec.breakerEnabled.configuration, "test", executionContext) {
-          override val jsonValidate: (JsLookupResult) => JsResult[String] = json => json.validate[String]
+
+          import AbstractService.implicitConversions._
+
+          override val validate: WSResponse ⇒ Try[String] = response ⇒ response.json.result.validate[String]
 
           override protected def initializeMetricsContext(name: String): Context = mockTimerContext
         }
@@ -40,16 +47,16 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         val service = new TestService()
         for {_ ← 1 to 100} {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
-        verify(mockRequest, times(100)).get()
+        verify(mockRequest, times(100)).execute(anyString())
       }
 
       "should reject requests when breaker opens" in new BreakerEnabled {
         val service = new TestService()
         service.breaker.fail()
         assertThrows[CircuitBreakerOpenException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
       }
 
@@ -57,7 +64,7 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
         val service = new TestService()
         service.breaker.fail()
         assertThrows[CircuitBreakerOpenException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
         // wait for breaker to half-open
         eventually(timeout(Span(2, Seconds)), interval(Span(25, Millis))) {
@@ -69,16 +76,16 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
         // next call should be OK
         when(responseMock.status).thenReturn(Helpers.OK)
         when(responseMock.json).thenReturn(JsString(""))
-        Await.result(service.get(), 1.second)
+        Await.result(service.execute(), 1.second)
 
-        verify(mockRequest).get()
+        verify(mockRequest).execute(anyString())
       }
 
       "open breaker if request fails when half-open" in new BreakerEnabled with Eventually {
         val service = new TestService()
         service.breaker.fail()
         assertThrows[CircuitBreakerOpenException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
         // wait for breaker to half-open
         eventually(timeout(Span(2, Seconds)), interval(Span(25, Millis))) {
@@ -89,10 +96,10 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
         service.breaker.isOpen mustBe true
         // next call should fail
         assertThrows[CircuitBreakerOpenException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
 
-        verify(mockRequest, never()).get()
+        verify(mockRequest, never()).execute(anyString())
       }
 
       "2xx errors will not open the breaker" in new BreakerEnabled {
@@ -101,7 +108,7 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
         when(responseMock.json).thenReturn(JsString("expected-value"))
 
         val service = new TestService()
-        Await.result(service.get(), 1.second) mustBe "expected-value"
+        Await.result(service.execute(), 1.second) mustBe "expected-value"
 
         service.breaker.isClosed mustBe true
       }
@@ -112,7 +119,7 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         val service = new TestService()
         assertThrows[HttpRedirectException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
 
         service.breaker.isClosed mustBe true
@@ -123,7 +130,7 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         val service = new TestService()
         assertThrows[HttpClientErrorException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
 
         service.breaker.isClosed mustBe true
@@ -135,11 +142,11 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
         val service = new TestService()
         // the first exception is returned and opens the breaker
         assertThrows[HttpServerErrorException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
         // the second request will hit the open breaker
         assertThrows[CircuitBreakerOpenException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
         service.breaker.isOpen mustBe true
       }
@@ -152,7 +159,10 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         class TestService extends AbstractService[String](mockWsClient, metricsMock,
           CircuitBreakerSpec.breakerDisabled.configuration, "test", executionContext) {
-          override val jsonValidate: (JsLookupResult) => JsResult[String] = json => json.validate[String]
+
+          import AbstractService.implicitConversions._
+
+          override val validate: WSResponse ⇒ Try[String] = response ⇒ response.json.result.validate[String]
 
           override protected def initializeMetricsContext(name: String): Context = mockTimerContext
         }
@@ -166,9 +176,9 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         val service = new TestService()
         for {i ← 1 to 100} {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
-        verify(mockRequest, times(100)).get()
+        verify(mockRequest, times(100)).execute(anyString())
       }
 
       "execute multiple requests even if failures occur" in new BreakerDisabled {
@@ -177,13 +187,13 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
         val service = new TestService()
 
         assertThrows[HttpClientErrorException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
         assertThrows[HttpClientErrorException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
 
-        verify(mockRequest, times(2)).get()
+        verify(mockRequest, times(2)).execute(anyString())
       }
 
       "2xx errors will return the expected content" in new BreakerDisabled {
@@ -192,7 +202,7 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
         when(responseMock.json).thenReturn(JsString("expected-value"))
 
         val service = new TestService()
-        Await.result(service.get(), 1.second) mustBe "expected-value"
+        Await.result(service.execute(), 1.second) mustBe "expected-value"
       }
 
       "3xx errors will cause redirect exceptions" in new BreakerDisabled {
@@ -201,7 +211,7 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         val service = new TestService()
         assertThrows[HttpRedirectException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
       }
 
@@ -210,7 +220,7 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         val service = new TestService()
         assertThrows[HttpClientErrorException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
       }
 
@@ -219,7 +229,7 @@ class CircuitBreakerSpec extends PlaySpec with MockitoSugar with TestExecutionCo
 
         val service = new TestService()
         assertThrows[HttpServerErrorException] {
-          Await.result(service.get(), 1.second)
+          Await.result(service.execute(), 1.second)
         }
 
       }