+++ pressed-client/src/main/scala/de/welt/contentapi/pressed/client/repository/PressedDiggerClient.scala
@@ -1,7 +1,6 @@
 package de.welt.contentapi.pressed.client.repository
 
 import javax.inject.{Inject, Singleton}
-
 import com.google.inject.ImplementedBy
 import com.kenshoo.play.metrics.Metrics
 import de.welt.contentapi.core.client.services.CapiExecutionContext
@@ -12,9 +11,10 @@ import de.welt.contentapi.pressed.models.PressedReads.apiPressedSectionResponseR
 import de.welt.contentapi.utils.Env.{Env, Live}
 import play.api.Configuration
 import play.api.libs.json.{JsLookupResult, JsResult}
-import play.api.libs.ws.WSClient
+import play.api.libs.ws.{WSClient, WSResponse}
 
 import scala.concurrent.Future
+import scala.util.Try
 
 @ImplementedBy(classOf[PressedDiggerClientImpl])
 sealed trait PressedDiggerClient {
@@ -31,12 +31,13 @@ sealed trait PressedDiggerClient {
 @Singleton
 class PressedDiggerClientImpl @Inject()(ws: WSClient, metrics: Metrics, configuration: Configuration, capi: CapiExecutionContext)
   extends AbstractService[ApiPressedSectionResponse](ws, metrics, configuration, "digger", capi) with PressedDiggerClient {
+  import AbstractService.implicitConversions._
 
-  override val jsonValidate: (JsLookupResult) ⇒ JsResult[ApiPressedSectionResponse] = _.validate[ApiPressedSectionResponse](apiPressedSectionResponseReads)
+  override val validate: WSResponse ⇒ Try[ApiPressedSectionResponse] = response ⇒ response.json.result.validate[ApiPressedSectionResponse]
 
   override protected[client] def findByPath(path: String, env: Env = Live)
                                            (implicit requestHeaders: RequestHeaders = Seq.empty): Future[ApiPressedSectionResponse] = {
-    get(urlArguments = Seq(env.toString, path), parameters = Nil)
+    execute(urlArguments = Seq(env.toString, path), parameters = Nil)
   }
 
 }